Update Configuration:
Convert IVaultProvider to IDisposable
Update current registered auto refresh providers to a dictionary to make concurrent removal easier.
Expose the current cached values to inheritied classes.
Convert mount and path to virutal to not introduce a breaking change by adding set to the mix.
Move the refresh thread from a simple thread wait to a wait event system to allow manual refreshing of all global providers.
Add argument to constructors to control if the provider should be added to the auto refresh thread or not.
Add extra constructor to allow setting mount and path via constructor.
Implement IDisposable